Feature/login player

+++ studentsystem/src/main/java/basketballleague/studentsystem/dto/TeamDTO.java
@@ -11,7 +11,7 @@ public class TeamDTO {
     private int id;
     private String name;
     private double totalPoints;
-    private double totalRebounds;
+    private float totalRebounds;
     private double totalAssists;
     private double totalSteals;
 }
